Register - shelter and user
